push_swap

+++ mandatory/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: adzahrao <adzahrao@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/04 23:04:37 by adzahrao          #+#    #+#             */
-/*   Updated: 2025/01/29 18:56:21 by adzahrao         ###   ########.fr       */
+/*   Updated: 2025/01/30 18:49:04 by adzahrao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ typedef struct s_push
 	int				flag;
 }					t_push;
 
-int					ft_ft_atoi(char *str);
+ssize_t				ft_ft_atoi(char *str);
 int					ft_atoi(char *str);
 int					check_repet(char *str, int *arr, int *size);
 int					h_error(char **argv, char *args);
